requirements.txt
ExifRead==3.0.0
pillow==11.0.0